+++ badminton-api/src/main/java/org/badminton/api/config/security/ClubPermissionEvaluator.java
@@ -2,7 +2,7 @@ package org.badminton.api.config.security;
 
 import java.io.Serializable;
 
-import org.badminton.api.interfaces.oauth.dto.CustomOAuth2Member;
+import org.badminton.api.interfaces.auth.dto.CustomOAuth2Member;
 import org.springframework.security.access.PermissionEvaluator;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
