+++ badminton-api/src/main/java/org/badminton/api/interfaces/auth/dto/GoogleResponse.java
@@ -1,0 +1,35 @@
+package org.badminton.api.interfaces.auth.dto;
+
+import java.util.Map;
+
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public class GoogleResponse implements OAuthResponse {
+
+	private final Map<String, Object> attribute;
+
+	@Override
+	public String getProvider() {
+		return "google";
+	}
+
+	@Override
+	public String getProviderId() {
+		return attribute.get("sub").toString();
+	}
+
+	@Override
+	public String getEmail() {
+		return attribute.get("email").toString();
+	}
+
+	@Override
+	public String getName() {
+		return attribute.get("name").toString();
+	}
+
+	public String getProfileImage() {
+		return attribute.get("picture").toString();
+	}
+}
