+++ badminton-api/src/main/java/org/badminton/api/interfaces/member/controller/MemberController.java
@@ -7,12 +7,12 @@ import org.badminton.api.aws.s3.model.dto.ImageUploadRequest;
 import org.badminton.api.aws.s3.service.MemberProfileImageService;
 import org.badminton.api.common.exception.member.ImageFileNotFoundException;
 import org.badminton.api.common.response.CommonResponse;
+import org.badminton.api.interfaces.auth.dto.CustomOAuth2Member;
 import org.badminton.api.interfaces.match.dto.MatchResultResponse;
 import org.badminton.api.interfaces.member.MemberDtoMapper;
 import org.badminton.api.interfaces.member.dto.MemberIsClubMemberResponse;
 import org.badminton.api.interfaces.member.dto.MemberMyPageResponse;
 import org.badminton.api.interfaces.member.dto.MemberUpdateRequest;
-import org.badminton.api.interfaces.oauth.dto.CustomOAuth2Member;
 import org.badminton.api.service.match.MatchResultService;
 import org.badminton.domain.domain.clubmember.info.ClubMemberMyPageInfo;
 import org.badminton.domain.domain.member.info.MemberIsClubMemberInfo;
