--- badminton-api/src/main/java/org/badminton/api/interfaces/member/dto/MemberDeleteResponse.java
@@ -1,20 +1,0 @@
-package org.badminton.api.interfaces.member.dto;
-
-import org.badminton.domain.domain.member.entity.Member;
-import org.badminton.domain.domain.member.info.MemberDeleteInfo;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-
-@Schema(description = "회원 삭제 responseDto")
-public record MemberDeleteResponse(
-        @Schema(description = "멤버 id", example = "1")
-        String memberToken,
-        @Schema(description = "삭제 여부", example = "true")
-        boolean isDeleted
-) {
-
-    public static MemberDeleteResponse fromMemberDeleteInfo(MemberDeleteInfo info) {
-        return new MemberDeleteResponse(info.memberToken(), info.isDeleted());
-    }
-
-}
