--- badminton-api/src/main/java/org/badminton/api/interfaces/oauth/dto/OAuthResponse.java
@@ -1,14 +1,0 @@
-package org.badminton.api.interfaces.oauth.dto;
-
-public interface OAuthResponse {
-
-    String getProvider(); // oAuth 제공자
-
-    String getProviderId(); // 제공자에서 발급해주는 아이디
-
-    String getEmail(); // 이메일
-
-    String getName(); // 설정한 이름
-
-    String getProfileImage();
-}
