+++ badminton-domain/src/main/java/org/badminton/domain/domain/auth/service/AuthService.java
@@ -1,14 +1,11 @@
 package org.badminton.domain.domain.auth.service;
 
-import org.badminton.domain.domain.member.info.MemberDeleteInfo;
+import org.badminton.domain.domain.auth.info.MemberDeleteInfo;
 
-import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
 public interface AuthService {
-	MemberDeleteInfo deleteMember(String accessToken, String registrationId);
+	MemberDeleteInfo deleteMember(String memberToken, String accessToken, String registrationId);
 
 	void logoutMember(String memberToken, HttpServletResponse response);
-
-	String refreshAccessToken(HttpServletRequest request, HttpServletResponse response);
 }
